requirements.txt
boto3==1.15.18
click==8.1.3
pydantic==1.9.1
cryptography==3.2.1
pydantic==1.9.1
pydantic-yaml==0.8.0
PyYAML==6.0.0
snowflake-snowpark-python==1.4.0
